refactor: 방문날짜 생성 로직 및 패키지 변경
